Change throttling behaviour to allow multiple calls
Instead of having a short throttle period for some commands, make the
throttle period longer but allow it to be called multiple times before
throttling.

Applying this to some commands where it makes sense for the command to
be called multiple times in quick succession because they have multiple
responses (like !quote or !advice) but not for commands that would only
give the same response (like !deathcount or !viewers).

Also, making !card available for non-mods to call as an experiment, with
a relatively long throttle. We'll see how it goes.