Make RPC connection more robust
If an exception happens on the bot side of the connection, pass the
exception back, so that the server side doesn't just end up waiting for
the socket to time out.

Pass across the traceback for easier debugging when using the Flask
debug mode, as it shows up on the error page.