Handle meld cards in mtgs scraper
Generate json in approximately the format I think is most likely for
mtgjson to use, once it has EMN cards. This will naturally need to be
revisted once the next mtgjson release happens.

The generated text for the chat is possibly a little verbose (especially
for Brisela and their super long card names) but I'm not sure how much
shorter they could be and still show the necessary info.