Add Embalm tokens
Can only be triggered by PubNub, not by `!card` in chat.

This is so that if someone says eg `!card angel sanction` then it still
just finds Angel of Sanctions, instead of saying "Did you mean Angel of
Sanctions or Angel of Sanctions token?"

In theory we could use this in future preprereleases for other special
cards that should be triggerable from the card reader, but would cause
confusion to be in `!card` proper.