Add "FRAMEWORK_ONLY" flag
Scripts can set this if they want to import lrrbot modules without
lrrbot.main automatically spinning up a bunch of the bot, and without
common.config monopolising the command line.

Lets us get rid of the sys.argv hacks in build_carddb, and the asyncio
warnings when that script ends.