Too-long-messages protection
Any PRIVMSG that gets sent will be trimmed to ensure it fits. Do the
trimming properly, to make sure it fits within the appropriate number of
bytes (rather than a number of chars).

Downside: we lose the processing from textwrap that cuts on words...
it'll cut in the middle of a word now. But I think we'll live.

Upside: we can shave off some of the just-to-be-safe buffer range, and
allow slightly longer messages.