Remove eris from the main project
The async branch of discord.py (the one that's on PyPI) isn't compatible
with Python 3.7 and it's dependencies are so old that we're getting
version conflicts. Thankfully the new versions of those dependencies seem
to be API compatible so we haven't had any crashes because of that yet.

The new version of discord.py, the rewrite branch is only available on
Github and I've heard rumors that there will never be a PyPI release. It
is technically possible for us to depend on a Git repository but I
really don't like doing it because it'd be a pain to keep up to date and
it'd expose us to broken intermediate states of that repository.

An alternative Python library exists (disco) but it doesn't seem to have
`asyncio` support and the documentation is a bit sparse.

After months of threatening to Rewrite It In Rust ('It' being eris) on
the #programming channel on Discord I managed to find some time to
finish it and I've decided to use the Rust version and retire the old
Python code.